requirements.txt
feedparser>=6.0.11
praw>=7.7.1
requests>=2.31.0
google-genai>=1.0.0
python-dotenv>=1.0.1
jinja2>=3.1.4
